[runtime_stats] Widen iteration counters, fix overhead total, test main_loop line
- Widen period/total iteration counters to uint64_t to avoid wrapping on
  long-running high-frequency loops.
- Compute total component-time sum over all components (not just the
  period-active subset) so total overhead is not inflated by components
  that ran earlier but are idle now.
- Extend integration test to parse the main_loop line and validate the
  iters/active_avg/active_total/overhead_total fields.

+++ esphome/components/runtime_stats/runtime_stats.h
@@ -53,11 +53,13 @@ class RuntimeStatsCollector {
   uint32_t log_interval_;
   uint32_t next_log_time_{0};
 
-  // Main loop active-time stats (wall time per iteration, excluding yield/sleep)
-  uint32_t period_active_count_{0};
+  // Main loop active-time stats (wall time per iteration, excluding yield/sleep).
+  // Counters are uint64_t — at sub-millisecond loop times a uint32_t can wrap in
+  // a few weeks of uptime, which is well within ESPHome device lifetimes.
+  uint64_t period_active_count_{0};
   uint64_t period_active_time_us_{0};
   uint32_t period_active_max_us_{0};
-  uint32_t total_active_count_{0};
+  uint64_t total_active_count_{0};
   uint64_t total_active_time_us_{0};
   uint32_t total_active_max_us_{0};
 };
